refactor: Add `_duration` module
Resolves (https://github.com/narwhals-dev/narwhals/pull/2465#discussion_r2072369830)
