+++ narwhals/_pandas_like/series_dt.py
@@ -4,12 +4,12 @@ from typing import TYPE_CHECKING
 from typing import Any
 
 from narwhals._compliant.any_namespace import DateTimeNamespace
+from narwhals._duration import parse_interval_string
 from narwhals._pandas_like.utils import PandasLikeSeriesNamespace
 from narwhals._pandas_like.utils import calculate_timestamp_date
 from narwhals._pandas_like.utils import calculate_timestamp_datetime
 from narwhals._pandas_like.utils import int_dtype_mapper
 from narwhals._pandas_like.utils import is_pyarrow_dtype_backend
-from narwhals.utils import parse_interval_string
 
 if TYPE_CHECKING:
     from narwhals._pandas_like.series import PandasLikeSeries
